Fibonacci Number Generator in Kotlin
This is a Kotlin Program to Generate Fibonacci Numbers. The number is
said to be in a Fibonacci series if each subsequent number is the sum
of the previous two numbers.

Enter the number of terms you want as an input. Now we use for loop to
generate the desired series.